Add auth to activity service

+++ backend/api-gateway/src/activity/activity.controller.ts
@@ -6,6 +6,8 @@ import {
   Param,
   Post,
   Put,
+  Req,
+  UseGuards,
 } from '@nestjs/common';
 import { ActivityService } from './activity.service';
 import { ApiTags } from '@nestjs/swagger';
@@ -16,36 +18,16 @@ import {
   JoinActivityDto,
   LeaveActivityDto,
 } from './activity.dto';
+import { JwtUserGuard } from 'src/guard/auth.guard';
 
 @ApiTags('Activities')
 @Controller('activities')
 export class ActivityController {
   constructor(private readonly activityService: ActivityService) {}
 
-  @Post()
-  createActivity(@Body() data: CreateActivityDto) {
-    console.log('data', data);
-    return this.activityService.createActivity(data);
-  }
-
   @Get()
   listActivities() {
     return this.activityService.listActivities();
-  }
-
-  @Put('/join')
-  joinActivity(@Body() data: JoinActivityDto) {
-    return this.activityService.joinActivity(data);
-  }
-
-  @Put('/leave')
-  leaveActivity(@Body() data: LeaveActivityDto) {
-    return this.activityService.leaveActivity(data);
-  }
-
-  @Put('/:id')
-  updateActivity(@Body() data: UpdateActivityDto, @Param('id') id: string) {
-    return this.activityService.updateActivity({ id, ...data });
   }
 
   @Get('/:id')
@@ -53,8 +35,47 @@ export class ActivityController {
     return this.activityService.getActivity({ id });
   }
 
+  @Post()
+  @UseGuards(JwtUserGuard)
+  createActivity(@Body() data: CreateActivityDto, @Req() req) {
+    if (!req.userId) throw new Error('User not found');
+    data.ownerId = req.userId;
+    return this.activityService.createActivity(data);
+  }
+
+  @Put('/join')
+  @UseGuards(JwtUserGuard)
+  joinActivity(@Body() data: JoinActivityDto, @Req() req) {
+    if (!req.userId) throw new Error('User not found');
+    data.userId = req.userId;
+    return this.activityService.joinActivity(data);
+  }
+
+  @Put('/leave')
+  @UseGuards(JwtUserGuard)
+  leaveActivity(@Body() data: LeaveActivityDto, @Req() req) {
+    if (!req.userId) throw new Error('User not found');
+    data.userId = req.userId;
+    return this.activityService.leaveActivity(data);
+  }
+
+  @Put('/:id')
+  @UseGuards(JwtUserGuard)
+  updateActivity(
+    @Body() data: UpdateActivityDto,
+    @Param('id') id: string,
+    @Req() req,
+  ) {
+    if (!req.userId) throw new Error('User not found');
+    data.ownerId = req.userId;
+    return this.activityService.updateActivity({ id, ...data });
+  }
+
   @Delete()
-  deleteActivity(@Body() data: DeleteActivityDto) {
+  @UseGuards(JwtUserGuard)
+  deleteActivity(@Body() data: DeleteActivityDto, @Req() req) {
+    if (!req.userId) throw new Error('User not found');
+    data.ownerId = req.userId;
     return this.activityService.deleteActivity(data);
   }
 }
